Move area line and point details files
