+++ packages/webapp/src/components/LocationDetailLayout/PureLocationDetailLayout.jsx
@@ -4,9 +4,9 @@ import { useForm, FormProvider } from 'react-hook-form';
 import LocationButtons from './LocationButtons';
 import LocationPageHeader from './LocationPageHeader';
 import Form from '../Form';
-import AreaDetails from './AreaDetails/AreaDetails';
-import LineDetails from './LineDetails/LineDetails';
-import PointDetails from './PointDetails/PointDetails';
+import AreaDetails from './AreaDetails';
+import LineDetails from './LineDetails';
+import PointDetails from './PointDetails';
 import RouterTab from '../RouterTab';
 import useLocationRouterTabs from '../../containers/LocationDetails/useLocationRouterTabs';
 import { Variant } from '../RouterTab/Tab';
